backend/core/requirements.txt
django-templated-mail==1.1.1
djoser==2.1.0
oauthlib==3.2.2
python3-openid==3.2.0
requests-oauthlib==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.4.2

asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
contourpy==1.0.6
cryptography==41.0.1
cycler==0.11.0
daphne==4.0.0
dj-database-url==2.0.0
Django==3.2.20
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
dnspython==2.3.0
fonttools==4.38.0
googlemaps==4.10.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
kiwisolver==1.4.4
matplotlib==3.5.3
msgpack==1.0.5
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pymongo==3.12.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
requests==2.31.0
scipy==1.7.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==22.10.0
## Esto tuve que sacarlo para poder desplegar en Render.com
# twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
zipp==3.15.0
zope.interface==6.0

boto3==1.28.7 
botocore==1.31.7 
django-storages==1.13.2 
jmespath==1.0.1 
s3transfer==0.6.1 
urllib3==1.26.16

whitenoise==6.5.0
